+++ template_annotation/lib/template_annotation.dart
@@ -1,4 +1,5 @@
 const template = Template();
+const constructor = Constructor();
 
 class Template {
   const Template({this.hiveType});
@@ -8,4 +9,9 @@ class Template {
 class Union {
   const Union(this.members);
   final Set<Type> members;
+}
+
+class Constructor {
+  const Constructor([this.name = '']);
+  final String name;
 }