+++ template_annotation/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.2.0
+- Added `Constructor` annotation for using static methods that create an
+  `Builder` update function as constructors.
+
 # 1.1.0
 - Added `hiveType` optional argument to the `Template` annotation